+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+
+- Extend env with client and server info
+
 ## [0.3.1] - 2025-10-08
 
 ### Fixed